Moving into existing connection package
